requirements.txt
flask~=3.1.0
psycopg2~=2.9.10
werkzeug~=3.1.3
MarkupSafe~=3.0.2
bcrypt~=4.2.1
PyJWT
